Fix overflows

+++ src/power/graphics/drawing/SelfContainedMeshCanvas.hpp
@@ -24,8 +24,8 @@ nanogui::Matrix4f glm_to_nanogui(glm::mat4 glmMatrix);
 class SelfContainedMeshCanvas : public nanogui::Canvas {
 private:
 	struct VertexIndexer {
-		size_t mVertexOffset = 0;
-		size_t mIndexOffset = 0;
+		unsigned int mVertexOffset = 0;
+		unsigned int mIndexOffset = 0;
 	};
 
 public:
